Add search mixin back and hook up community api urls
