--- backend/api/mixins/__init__.py
@@ -1,1 +1,0 @@
-from .search_mixin import SearchMixin as SearchMixin
