+++ backend/apps/api/urls.py
@@ -3,4 +3,5 @@ from django.urls import path, include
 urlpatterns = [
     path("auth/", include(("apps.authentication.urls", "authentication"))),
     path("users/", include(("apps.users.urls", "users"))),
+    path("communities/", include(("apps.communities.urls", "communities"))),
 ]
